moved Tags methods to TagSelect (renamed) class
